fm: defer imbalance check to the end of passes
With f113789a1bd1eb89a0cc4f40c368a64c6f07f4fd fidducia-mattheyses can
restrict its movements to a given imbalance tolerance, some solutions
become inaccessible with the default settings (when fm does not go above
the input partition's imbalance).

This is the case of the doc example of the algorithm, and the reason why
a "0.25" threshold has been chosen (which is actually (5/8-1/2)/(1/2),
the imbalance of the partition after the first move of the algorithm).

By checking the imbalance once the pass have finished, FM can obtain the
desired partition without allowing imbalances up to 25% (which is a
relatively weak tolerance in the real world).