tools: add medit2svg
Converts a medit file (.mesh) to a somewhat lightweight SVG file.

We want a SVG following this format:

    <svg viewBox="xmin ymin width height" xmlns="???">
    <path d="..." fill="#abcdef"/>
    <path d="..." fill="#abcdef"/>
    ...
    </svg>

Each path's "d" attribute is a list of commands (M for move, L for line,
Z to close and fill) that draw a path, for example:

    <path d="M 0 0 L 0 10 L 10 10 L 10 0 Z" fill="red" />

draws a red square between (0,0) and (10,10), and

    <path d="M 0 0 L 10 0 L 0 10 Z" fill="black" />

draws half of this square in black.

A naive approach to convert mesh files to SVGs is simply to make a path
for each cell.  It works, but consumes both a lot of space and a lot of
processing power to render it.

So I'm trying to figure out how to get a path for each equivalence
class.  Turns out there are many edges cases and is a lot more tedious
than I expected...

Finding out equivalence classes is easy, and so is finding cells at the
frontier of these classes.  The problem is converting the set of cells
in the frontier to a list of nodes that make up a path.

Some problems I encountered:

1. the path must be ordered, otherwise it will not render well,
2. there can be nested equivalence classes, and they will therefore lead
   to 2 or more paths, one for the outside and one for each nested
   equivalence class,
3. some nodes can appear *twice* in a path, for example see this quad
   mesh with two classes `X` and `.`:

       ....
       ..X.
       XX..  <-- the . on the left of this row, its upper left node
       XXXX      appears twice on `.`'s path, because it's also a node
                 of the upper left `.`

The solution to nested classes turns out to be built in SVG itself: by
setting a "fill-rule" attribute[0] to "evenodd" and drawing multiple shapes
in a path, it's possible to "carve out" portions of the class.

The solution to nodes that appear twice in a path is to first merges the
edges of frontier elements into a path, and then merging these paths
together.  I don't have a formal proof, but it seems that with this
first pass, the problem disappears.

[0] https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/fill-rule