Add a C FFI to coupe
So that coupe can be used from most languages, not just rust.

The bindings uses `uintptr_t` extensively, since it's the equivalent of
`usize` (and `size_t` is not[0]).

The C interface follows this pattern:

A `coupe_data` opaque type can be constructed in several ways, to allow
the user to feed coupe elements in different ways.

For example, if the user has an array of weights, they can use:

    double weights[4] = {1.2, 3.4, 5.6, 7.8};
    coupe_data *w =
        coupe_data_array(4, COUPE_DOUBLE, weights);

or if they don't want to use weights:

    int one = 1;
    coupe_data *w = coupe_data_constant(weight_count, COUPE_INT, &one);

(very idiomatic C!  unfortunately for these function to be "generic" the
second argument has to be a `void *`)

and if they have a complex data structure that can be indexed:

    ComplexWeightCollection *weights = /* ... */
    void *coll_get_ith(ComplexWeightCollection *coll, uintptr_t i) {
        /* ... */
    }
    coupe_data *w = coupe_data_fn(weights, weight_count, coll_get_ith);

For topologic algorithms, a `coupe_adjncy` type is defined.  For now it
can only be constructed from compressed adjacency matrices.

Algorithms are implemented as functions that take:

- the partition array,
- some `coupe_data` for the points and weights,
- some `coupe_adjncy` for the topology,
- the algorithms parameters.

They return an error if any.  Maybe some error can be returned but the
partition still be populated and valid.

[0] https://rust-lang.github.io/unsafe-code-guidelines/layout/scalars.html#isize-and-usize