ffi: do not document rust crate
We want coupe-ffi to have a library named coupe, so that the generated
objects are libcoupe.{a,so}.  However, this makes `cargo doc` fail
because it conflicts with the rust lib.

Since the rust-side of the documentation is not important (there already
is a documented, doxygen-compatible header file), this commit disables
the documentation on this crate.