rcb: remove async implementation
for simplicity's sake

Zoltan2 uses the same BSP paradigm + data movements, so the simple_rcb
version is state of the art.  A next step would be to improve data
locality with custom rayon drivers.

This commit removes the relevant dependencies, and also removes the
indirect dependency on proc-macros via nalgebra-macros and tracing.