rcb: array of structs -> struct of arrays
Unpack the array of Item<D> into:

- an array for the weights,
- an array for the part references,
- D arrays, one for each coordinates

Also, change the point type from f64 to f32.

All this is to allow efficient data reads and copies, so that SIMD and
GPU acceleration can be used.  This change alone increases performance,
maybe because of the lower amount of data that needs to be read.

This commit also replaces the use of select_nth_unstable with a variant
of the stdlib "partition" algorithm:

<https://github.com/rust-lang/rust/blob/1d6010816c37186e2bee316709f0c0197c427513/library/core/src/slice/sort.rs#L544>

(basically the same thing as select_nth_unstable, but the pivot is given
as input so there's less computation to be done)