Enable the rust_2018_idioms lint as warning
This should become the default in the future:
https://github.com/rust-lang/rust/issues/54910

Reason this is put in lib.rs instead of CI: to ensure local invocations
of cargo build and cargo clippy pick up the warning.

Reason this is made a warning: so that code builds and runs anyway. On
the CI, the `-D warnings` clippy argument will treat it as a hard error.