arcswap: fix imbalance violations
Several changes:

1. arcswap now works in passes,
2. the way `max_part_weight` works

Change #1 is a workaround the fact that arcswap was exiting early
instead of doing moves until there is none. There might be more
efficient ways of doing this, since this is mostly bulk synchronous.

It has however the advantage of easing the thread-local imbalance
constraints, since they are now updated before each next pass.

This brings us to change #2. The old ways of `max_part_weight` might not
have been working at all, even in the bi-partitioning case. Now,
`max_part_weight` is the real (as in not thread-local) maximum weight a
part can have without breaking the imbalance constraint. Thread-local
constraints are now specified for each part as a read-only array of the
maximum weight each part can have. It is set such that if all threads
move as much weight as they can to one specific part, then this part
will still be under the imbalance threshold.

Does it work? In most cases, yes. There is however some 1% of the time
where it slightly goes above the threshold.