hilbert: add a simple 3D implementation
The 3D implementation needs a smaller lookup table, because there are
twice as many configurations to go from an order to the next. Therefore,
the LUT allows for only 3 bit to be computed at a time (instead of 6).
Its layout is similar however, configurations are encoded on 3 bits
instead of 2.

To prevent trait implementation conflicts, HilbertCurve can only
implement Partition for &[PointND] instead of all P: AsRef<[_]>.