report: make efficiency print wall-time graphs too
Now, the efficiency script directly outputs the timings to .wt.csv.
Then we use awk to process this csv file and compute another one with
the speedups (awk is actually really useful to process csv files).

We don't use log scales for the strong scaling graph, to avoid focus on
low core counts.