weighted_quantiles: fix splits "tearing"
Previously, weighted_quantiles would modify "splits" in place. While no
actual bug or imbalance violations have been observed, this caused
"splits" and "part_weights" to go out of sync. Later iterations of the
inner for loop would use the newer splits with the former values of
part_weights.

The split update is not done in parallel, because the overhead of rayon
becomes too costly with high core counts.

It seems that this version also scales a bit better than master's, maybe
because the for loop has been replaced with a closure and leads to
better behavior wrt. the instruction cache? We'll never know.