rcb for cartesian meshes
With cartesian meshes (grids), we have some additional properties:

- weights are sorted,
- points are grid indices from 0 to a given N.

The weighted_median procedure could simply loop through the weights and
take the best value for a cut, that would be O(n). In our case, we use
fold_chunks to compute in parallel the weight of T portions of the axis,
find the chunk where the cut lies and then use fold_chunks on that chunk
alone. I believe that makes it O(n/t * log_t(n)), which should(?) be
better (if log_t(n)/t < 1 ?).

The fold_chunks implementation has been taken from a PR to rayon:
https://github.com/rayon-rs/rayon/pull/970