cartesian/rcb: start iterations at coord=1
In both 2D and 3D cases, it makes the weight iterator given to
weighted_median iterate through the X axis in the inner-most loop. This
gives the benefit of cache predictions on the first iteration where we
have few tasks to give to rayon. This leads to a 15% perf increase.