tools: fix integer scaling for metis/scotch
The idea is that we have weights
- in some [min; max] range,
- whose sum is S

and we want to scale them so
1. they fit into some integer type,
2. the S is below INTMAX after scaling
3. the output has no zeroes

basically we apply an affine function "y=scale*x+offset" where
scale and offset depend on the max output value M (S, scaled):

         M * (x-min) + max-x
    y = ---------------------
               max-min

y=1 for x=min and y=M for x=max

M is set so that the sum of all "y" is INTMAX:

         (max-min) * INTMAX - sum(max-x)
    M = ---------------------------------
                   sum(x-min)

However, with this method, [0,1,2,3] and [10000,10001,10002,10003] are
scaled the same. In the context of load balancing, we would like the
later array to be scaled differently. Typically, the output values
should be closed to each other. On the contrary, the values of the
former array should be scattered accross the INT range. For this reason,
we set min=0 in the above calculations (if min is positive).